# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bandjacks API",
    "description": "\n    ## Cyber Threat Defense World Modeling API\n    \n    Bandjacks provides comprehensive cyber threat intelligence operations including:\n    \n    - **ATT&CK Data Management**: Load and manage MITRE ATT&CK releases\n    - **Natural Language Search**: Hybrid vector and graph search for CTI\n    - **Graph Exploration**: Traverse and analyze threat relationships\n    - **Document Extraction**: Extract CTI entities from documents using LLMs\n    - **Review Workflows**: Human-in-the-loop validation and feedback\n    - **Defense Recommendations**: D3FEND integration (coming in Sprint 4)\n    \n    ### Key Features\n    \n    - STIX 2.1 compliant data model\n    - ADM (ATT&CK Data Model) validation\n    - Hybrid search with vector embeddings and graph patterns\n    - LLM-powered extraction with evidence grounding\n    - Review queue for candidate validation\n    - Performance optimized with caching and connection pooling\n    \n    ### Authentication\n    \n    Currently no authentication required (development mode).\n    Production deployments should implement appropriate authentication.\n    \n    ### Rate Limiting\n    \n    No rate limiting currently implemented.\n    Production deployments should add rate limiting.\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Bandjacks Team",
      "email": "support@bandjacks.io"
    },
    "license": {
      "name": "Apache 2.0",
...
    }
...
  "servers": [
...
  "paths": {
...
    "/v1/feedback/relevance": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit Relevance Feedback",
        "description": "Submit relevance feedback for search results.\n    \n    This feedback is used to:\n    - Improve search ranking algorithms\n    - Learn query patterns and user preferences\n    - Identify gaps in the knowledge base\n    - Train machine learning models for better matching\n    \n    Feedback is stored and aggregated for continuous improvement.",
        "operationId": "submit_relevance_feedback_v1_feedback_relevance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelevanceFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Feedback successfully recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid feedback data"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feedback/correction": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit Correction Feedback",
        "description": "Submit a correction for an object or relationship.\n\nCorrections require review before being applied to maintain\ndata quality and audit trail.",
        "operationId": "submit_correction_feedback_v1_feedback_correction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feedback/validation": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit Validation Feedback",
        "description": "Submit validation feedback for extracted entities.\n\nThis is used to approve or reject candidates from extraction\nbefore they are merged into the main knowledge graph.",
        "operationId": "submit_validation_feedback_v1_feedback_validation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feedback/pending": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "Get Pending Feedback",
        "description": "Get pending feedback items that need review or action.\n\nReturns corrections pending review and other actionable feedback.",
        "operationId": "get_pending_feedback_v1_feedback_pending_get",
        "parameters": [
          {
            "name": "feedback_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feedback Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Pending Feedback V1 Feedback Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/feedback/{feedback_id}/apply": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Apply Feedback",
        "description": "Apply approved feedback (corrections).\n\nThis actually modifies the knowledge graph based on the feedback.\nRequires appropriate permissions.",
        "operationId": "apply_feedback_v1_feedback__feedback_id__apply_post",
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feedback Id"
            }
          },
          {
            "name": "reviewer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reviewer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Apply Feedback V1 Feedback  Feedback Id  Apply Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorrectionFeedback": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id",
            "description": "STIX ID to correct"
          },
          "correction_type": {
            "type": "string",
            "enum": [
              "property",
              "relationship",
              "classification"
            ],
            "title": "Correction Type",
            "description": "Type of correction"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "Field to correct (for property corrections)"
          },
          "old_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Old Value",
            "description": "Current value"
          },
          "new_value": {
            "title": "New Value",
            "description": "Proposed new value"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "description": "Reason for correction"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence",
            "description": "Supporting evidence"
          },
          "analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Id",
            "description": "Analyst identifier"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "correction_type",
          "new_value",
          "rationale"
        ],
        "title": "CorrectionFeedback",
        "description": "Correction feedback for objects or relationships."
      },
...
      "FeedbackResponse": {
        "properties": {
          "feedback_id": {
            "type": "string",
            "title": "Feedback Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "feedback_id",
          "status",
          "message"
        ],
        "title": "FeedbackResponse",
        "description": "Response for feedback submission."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RelevanceFeedback": {
        "properties": {
          "query_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Id",
            "description": "Associated query ID"
          },
          "object_id": {
            "type": "string",
            "title": "Object Id",
            "description": "STIX ID of the object"
          },
          "relevance": {
            "type": "string",
            "enum": [
              "relevant",
              "irrelevant",
              "needs_context"
            ],
            "title": "Relevance",
            "description": "Relevance judgment"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Optional comment"
          },
          "analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Id",
            "description": "Analyst identifier"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "relevance"
        ],
        "title": "RelevanceFeedback",
        "description": "Relevance feedback for search results."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationFeedback": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id",
            "description": "STIX ID to validate"
          },
          "validation": {
            "type": "string",
            "enum": [
              "approve",
              "reject",
              "needs_review"
            ],
            "title": "Validation",
            "description": "Validation decision"
          },
          "confidence_adjustment": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": -50.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Adjustment",
            "description": "Confidence adjustment (-50 to +50)"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Reason for decision"
          },
          "analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Id",
            "description": "Analyst identifier"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "validation"
        ],
        "title": "ValidationFeedback",
        "description": "Validation feedback for extracted entities."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication (future implementation)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication (future implementation)"
      }
    },
...
  },
...
}